apps/ai-orchestrator-crewai/pyproject.toml
[project]
name = "ai-orchestrator-crewai"
version = "0.1.0"
description = "Isolated CrewAI pilot service for EduAssist comparison runs."
requires-python = ">=3.12"
dependencies = [
  "crewai[google-genai]==1.12.2",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.8.0",
  "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_orchestrator_crewai"]
